refactor(query-performance-ai): split orchestrator vs sandboxed agent
Reorganises the product so that host-side orchestration code and
sandbox-side agent code live in separate top-level subdirs, and
test files sit next to the code they exercise.

  orchestrator/
    apps.py, harvest.py, coordinator.py, server.py, slow_queries.py
    backends/{base,local,metabase}.py
    backends/tests/test_backends.py
    tests/{test_server,test_slow_queries}.py
  sandboxed_autoresearch_agent/
    driver/{run_campaign,runtime,sandboxing}.py     # host-side driver
    pi_plugin/                                      # pi install-able
      package.json, docs/, prompts/, skills/, templates/
      scripts/{_common,ch_*,transports}.py
      scripts/tests/test_transports.py

INSTALLED_APPS path updated; Django app verified to load. All 29
tests pass; mypy clean across all 25 source files. The Python
package path also changed for python -m:
  products.query_performance_ai.scripts.coordinator
    -> products.query_performance_ai.orchestrator.coordinator
  products.query_performance_ai.scripts.run_campaign
    -> products.query_performance_ai.sandboxed_autoresearch_agent.driver.run_campaign

Generated-By: PostHog Code
Task-Id: 9ee4b87f-fce9-4d82-aded-510477a12e4a
