--- products/query_performance_ai/autoresearch/templates/workspace/adapter.json.example
@@ -1,6 +1,0 @@
-{
-  "_comment": "Coordinator transport. The only supported `type` is 'coordinator' (see products/query_performance_ai/autoresearch/scripts/transports.py). `url` is the coordinator's HTTP server (host.docker.internal from inside a Docker sandbox); `token` is the random per-run token the coordinator prints to its stdout at startup.",
-  "type": "coordinator",
-  "url": "http://host.docker.internal:8765",
-  "token": "<coordinator token>"
-}
