+++ products/query_performance_ai/orchestrator/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class QueryPerformanceAiConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "products.query_performance_ai.backend"
+    name = "products.query_performance_ai.orchestrator"
     label = "query_performance_ai"
