+++ products/query_performance_ai/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@posthog/products-query-performance-ai",
     "scripts": {
-        "backend:test": "pytest products/query_performance_ai/backend/tests"
+        "backend:test": "pytest products/query_performance_ai/orchestrator products/query_performance_ai/sandboxed_autoresearch_agent"
     }
 }
