+++ products/query_performance_ai/sandboxed_autoresearch_agent/pi_plugin/templates/workspace/campaign.json.example
